+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Performance tests now use reduced dataset size in CI to avoid timeouts
+
 ## [3.1.16] - 2026-01-21
 
 ### Added